==> process/03_Load.ipynb <==
# ETL Faits_Matières

Ce notebook extrait, transforme et charge les données de matières résiduelles dans la table de fait principale du projet BI.

import pandas as pd
import os
os.makedirs('../generated', exist_ok=True)
source_dir = '../generated/normalized/'




# # Extraction des données
# faits = pd.read_csv(f"{source_dir}matieres_residuelles_quebec.csv", sep=',', encoding='utf-8')

# # Transformation : sélection et renommage des colonnes
# faits_matieres = faits[[
#     'PÉRIODE DE RÉFÉRENCE',
#     'GÉO',
#     'Type de matières récupérées',
#     'Sources des matières préparées en vue du recyclage',
#     'VALEUR'
# ]].rename(columns={
#     'PÉRIODE DE RÉFÉRENCE': 'periode',
#     'GÉO': 'geo',
#     'Type de matières récupérées': 'type_matiere',
#     'Sources des matières préparées en vue du recyclage': 'producteur',
#     'VALEUR': 'volume_tonnes'
# })

# faits_matieres.head()

# # Sauvegarder le fichier pour l'analyse exploratoire
# faits_matieres.to_csv('../generated/faits_matieres.csv', index=False)

# Extraction des données massiques (fait_massique)
faits_matieres = pd.read_csv(f"{source_dir}matieres_residuelles_massique.csv", sep=',', encoding='utf-8')


# Affichage des premières lignes pour vérification
display(faits_matieres.head())

# Sauvegarder le fichier pour l'analyse exploratoire (optionnel)
faits_matieres.to_csv('../generated/faits_matieres.csv', index=False)


faits = pd.read_csv(f"{source_dir}matieres_residuelles_massique.csv", sep=',', encoding='utf-8')

sites = pd.read_csv(f"{source_dir}liste-centres-tri-crd.csv" , sep=',', encoding='utf-8')

# Dimension TypeMatiere
dim_type_matiere = pd.DataFrame({'type_matiere': faits['matiere'].unique()})
# Dimension Producteur
dim_producteur = pd.DataFrame({'producteur': faits['territoire'].unique()})

# Dimension Temps
dim_temps = pd.DataFrame({'periode': faits['annee'].unique()})

# Dimension SiteValorisation
dim_site_valorisation = sites.drop_duplicates().reset_index(drop=True)

print("Dimension TypeMatiere")
display(dim_type_matiere.head())

print("\nDimension Producteur")
display(dim_producteur.head())

print("\nDimension Temps")
display(dim_temps.head())

print("\nDimension SiteValorisation")
display(dim_site_valorisation.head())




# Persistance des dimensions en CSV
dim_type_matiere.to_csv('../generated/dim_type_matiere.csv', index=False)
dim_producteur.to_csv('../generated/dim_producteur.csv', index=False)
dim_temps.to_csv('../generated/dim_temps.csv', index=False)
dim_site_valorisation.to_csv('../generated/dim_site_valorisation.csv', index=False)
print("Dimensions sauvegardées dans des fichiers CSV à la racine du projet.")
