==> process/04_Analyse_Exploratoire.ipynb <==
# Analyse exploratoire

Ce notebook permet d'explorer les données de faits et de dimensions pour en vérifier la qualité et en extraire des statistiques descriptives.

import pandas as pd
source_dir = '../generated/'
faits_matieres = pd.read_csv(f"{source_dir}faits_matieres.csv", sep=',', encoding='utf-8')

# Statistiques descriptives
faits_matieres.describe()

# Répartition par type de matière
faits_matieres['matiere'].value_counts()



# Nombre de valeurs manquantes par colonne
print("Valeurs manquantes par colonne :")
print(faits_matieres.isnull().sum())

# Statistiques sur la colonne 'quantite_generee_donnees_agglo'
print("\nStatistiques sur 'quantite_generee_donnees_agglo' :")
print(faits_matieres['quantite_generee_donnees_agglo'].describe())

print("\nStatistiques sur 'quantite_collectee_donnees_agglo' :")
print(faits_matieres['quantite_collectee_donnees_agglo'].describe())


# Répartition des données par année ('annee')
print("\nRépartition par année :")
print(faits_matieres['annee'].value_counts().sort_index())

# Top 5 des types de matières les plus fréquents
print("\nTop 5 des types de matières :")
print(faits_matieres['matiere'].value_counts().head())

# # Top 5 des producteurs les plus fréquents
# print("\nTop 5 des producteurs :")
# print(faits_matieres['producteur'].value_counts().head())

# Répartition géographique
print("\nRépartition géographique :")
print(faits_matieres['territoire'].value_counts().head())
